+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- PNDA-3391: add more metrics to kafka plugin which is now managed in a by a JMX config file
+
 ### Changed
 - PNDA-3601: disable emailtext in Jenkins file and replace it with notifier stage and job